Chore: types 분리 파일 제거
