+++ FE/src/components/Modal/Potal.tsx
@@ -1,7 +1,13 @@
+import { ReactNode } from 'react';
 import { createPortal } from 'react-dom';
 
-import { PortalProps } from '@/components/Modal/types';
 import useScrollLock from '@/hooks/useScrollLock';
+
+// TODO: 일단 가장 넓은 범위의 ReactNode 타입 사용 (추가 공부 필요)
+export interface PortalProps {
+  parent: ReactNode | null | undefined;
+  children: ReactNode;
+}
 
 export default function Potal({ parent, children }: PortalProps) {
   useScrollLock(true); // TODO: advanced 버전으로 변경해보기
